extracted payload comparison methodss to their own class
