+++ testing/integration/test.py
@@ -4,10 +4,13 @@ import requests
 import json
 import sys
 sys.path.append('../../taketwo-webapi')
+sys.path.append('../util')
 from main import app
 from main import validate
 from main import db
 from fastapi.testclient import TestClient
+from testutil import comparePayloads
+import unittest
 
 def override_validate():
     return {"sub": 'test'}
@@ -38,16 +41,6 @@ def test_save_mark():
     
     # print response full body as text
     print(resp.text)
-    
-def comparePayloads(expected: dict, actual: dict):
-    assert expected["user_id"] == actual["user_id"], showDiff(str(expected["user_id"]), str(actual["user_id"]))
-    assert expected["flagged_string"] == actual["flagged_string"], showDiff(str(expected["flagged_string"]), str(actual["flagged_string"]))
-    assert expected["category"] == actual["category"], showDiff(str(expected["category"]), str(actual["category"]))
-    assert expected["info"] == actual["info"], showDiff(str(expected["info"]), str(actual["info"]))
-    assert expected["url"] == actual["url"], showDiff(str(expected["url"]), str(actual["url"]))
-
-def showDiff(expected: str, actual: str):
-    return "Expected: " + expected + ", Actual: " + actual
 
 if __name__ == '__main__':
     test_save_mark()
